--- app/src/main/java/io/github/jangerhard/BitcoinWalletTracker/client/BlockonomicsRequest.java
@@ -1,4 +1,0 @@
-package io.github.jangerhard.BitcoinWalletTracker.client;
-
-public class BlockonomicsRequest {
-}
