+++ app/src/main/java/io/github/jangerhard/BitcoinWalletTracker/model/BlockonomicsRequest.java
@@ -1,0 +1,4 @@
+package io.github.jangerhard.BitcoinWalletTracker.model;
+
+public class BlockonomicsRequest {
+}
