+++ app/src/main/java/io/github/jangerhard/BitcoinWalletTracker/model/BlockonomicsResponse.java
@@ -1,0 +1,6 @@
+package io.github.jangerhard.BitcoinWalletTracker.model;
+
+public class BlockonomicsResponse {
+
+
+}
