+++ app/src/main/java/io/github/jangerhard/BitcoinWalletTracker/utilities/BitcoinUtils.java
@@ -11,7 +11,11 @@ import java.util.Locale;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import com.samourai.wallet.util.FormatsUtilGeneric;
-import io.github.jangerhard.BitcoinWalletTracker.client.BlockinfoResponse;
+import io.github.jangerhard.BitcoinWalletTracker.model.BlockinfoResponse;
+import io.github.jangerhard.BitcoinWalletTracker.model.Transaction;
+import io.github.jangerhard.BitcoinWalletTracker.model.TransactionInput;
+import io.github.jangerhard.BitcoinWalletTracker.model.TransactionOut;
+import io.github.jangerhard.BitcoinWalletTracker.model.TransactionPrevOut;
 import io.vavr.collection.List;
 import io.vavr.control.Option;
 import org.bitcoinj.params.MainNetParams;
