+++ app/src/test/java/io/github/jangerhard/BitcoinWalletTracker/utilities/BitcoinUtilsTest.java
@@ -2,7 +2,11 @@ package io.github.jangerhard.BitcoinWalletTracker.utilities;
 
 import java.util.ArrayList;
 
-import io.github.jangerhard.BitcoinWalletTracker.client.BlockinfoResponse;
+import io.github.jangerhard.BitcoinWalletTracker.model.BlockinfoResponse;
+import io.github.jangerhard.BitcoinWalletTracker.model.Transaction;
+import io.github.jangerhard.BitcoinWalletTracker.model.TransactionInput;
+import io.github.jangerhard.BitcoinWalletTracker.model.TransactionOut;
+import io.github.jangerhard.BitcoinWalletTracker.model.TransactionPrevOut;
 import io.vavr.collection.List;
 import io.vavr.control.Option;
 import org.junit.Before;
